requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
selenium==4.16.0
webdriver-manager==4.0.1
Pillow==10.2.0
pydantic==2.5.3
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
httpx==0.26.0
mysql-connector-python==8.3.0
